MySQL issues resolved
Good grief. I think I've finally got the database issues figured out. My user was getting created with the % host parameter, but I was trying to grant privileges on the localhost host, so it did nothing. I'm specifying the user to be created with localhost as the host and granting privileges with localhost also.

Signed-off-by: Chris Wilcoxson <chris@slushman.com>