Corrected new site error
Newsite.sh had an error in the new site with WordPress section that didn't replace the placeholder text in the defaultsite.settings file when a new site was created. It was looking for a file with the full site domain in the name, rather than just the site title (christopherjoel.com vs christopherjoel).

Signed-off-by: Chris Wilcoxson <chris@slushman.com>